HomeHeader, ShareButton 컴포저블 함수로 분리
